refactor: eliminate account/ barrel, route consumers to sibling files
account/index.ts was a barrel with:

    export * as Account from "./account"
    export { AccountID, AccessToken, RefreshToken, ... } from "./schema"
    export type { AccountOrgs, ActiveOrg } from "./account"

This is the "barrel with many sibling passthroughs" shape the spec
explicitly calls out as bad: importing anything from @/account eagerly
evaluates BOTH ./schema AND ./account (which pulls in HTTP clients,
Effect layers, etc.), even if the consumer only wanted a schema type.

Per the spec's "When NOT to collapse to index.ts" section, account/
has multiple independent siblings (account, schema, repo, url, *.sql)
and should not be collapsed. Instead, eliminate the barrel entirely
and route consumers to the specific sibling files:

  - Delete account/index.ts
  - Append `export * as Account from "./account"` to account/account.ts
    so `import { Account } from "@/account/account"` works
  - Update 8 consumer files to import from siblings:
    * Account namespace \u2192 @/account/account
    * AccountID, AccessToken, etc. \u2192 @/account/schema
    * Mixed imports split into two lines

Dead code removed: `export type { AccountOrgs, ActiveOrg }` was never
imported anywhere (verified via rg).

Verified: tsgo --noEmit 0 errors, SDK generate no drift,
test/account/ + share/share-next + config/config 106/106 pass.
