+++ packages/opencode/src/account/account.ts
@@ -452,3 +452,5 @@ export const layer: Layer.Layer<Service, never, AccountRepo.Service | HttpClient
 )
 
 export const defaultLayer = layer.pipe(Layer.provide(AccountRepo.layer), Layer.provide(FetchHttpClient.layer))
+
+export * as Account from "./account"
