+++ packages/opencode/src/server/instance/experimental.ts
@@ -10,7 +10,8 @@ import { MCP } from "../../mcp"
 import { Session } from "../../session"
 import { Config } from "../../config"
 import { ConsoleState } from "../../config/console-state"
-import { Account, AccountID, OrgID } from "../../account"
+import { Account } from "../../account/account"
+import { AccountID, OrgID } from "../../account/schema"
 import { AppRuntime } from "../../effect/app-runtime"
 import { errors } from "../error"
 import { lazy } from "../../util/lazy"
